tests: rename fetchers to adapters
